Updated git submodules
Project: genius master 9a790c1be8b35218188d324f0ccbbb34a9924cfa

Fix for the Bug 5533 - NumberFormatException in AlivenessProtocolHandlerLL.java

Problem Statement : The NumberFormatException was observed in VPN service if 3 node cluster are started at the same time in the class AlivenessProtocolHandlerLL.java. This was reported in vpnservice and genius.

Analysis :  By looking at the implemenation of AlivenessProtocolHandlerLLDP.java, there is a unused local variable (i.e. packetId ) in handlePacketIn() mentod which caused the issue. This will be definitly fixed if the assignment is removed since the

Fix Detail: The unused variable (i.e. packetId ) in handlePacketIn() from the class AlivenessProtocolHandlerLLDP.java is removed to avoid this issue

Change-Id: Idbca6a6ba31626842ded7d82f00e9693a0dcbf5f
Signed-off-by: Dhevendran <dhevendran.k@tcs.com>