Updated git submodules
Project: yangtools master c8e8e94e8f1d398e9afe6b585ac9060e95ec7daa

Deprecate {String,UUID}Identifier

These identifiers are easigy mixed up if we do not allow for subclassing. Deprecate
the classes and provide their abstract equivalents. This allows for specific subclasses
which do not produce clashes.

Change-Id: I7a23a3e1180ee2bbfbc9997bec3ffee9c0db297f
Signed-off-by: Robert Varga <rovarga@cisco.com>


Annotate Identifier with @ThreadSafe

Implementations are expected to be used for maps, hence they are expected
to be used from multiple threads concurrently.

Change-Id: I7578486cc8be843959befb4473461dcd601e904b
Signed-off-by: Robert Varga <rovarga@cisco.com>


Introduce WritableObject/WritableIdentifier

When serializing nested Identifiers, it is sometimes useful
to be able to skip the object header in the stream. This patch
introduces WritableObject interface and WritableIdentifier
as the unification of WritableObject and WritableIdentifier.

It furthermore adds utilities to write long values if a format
which uses minimum number of bytes.

Change-Id: I8459e2883b4ab3dff113000cea43a452eacd3324
Signed-off-by: Robert Varga <rovarga@cisco.com>