Updated git submodules
Project: ovsdb stable/beryllium 941971288b13e5bc1af6a8fee19358e36ed9b404

Bug 5992 - GBP + SFC integration is broken

When Bridge and termination points are created together, it generates two data tree change
notifications. Termination point changes are part of second data change notification,
but that does not contain bridge and node augmentation, because there was no change
in those augmentations, and because of that plugin was not able to find the connection
instance for the ovsdb node. As a last resort plugin tries to find it based on the
rootNode iid from operational data store, but because in this scenario rootNode is
bridge node, and bridge creation request was process just before the second data
change notification, operational data was not updated with the bridge node details
by the time plugin checks in the operational data store. So final attempt to find
the connection instance also fails. To fix this and also optimize it a bit, this
patch added an additional attempt to get the connection instances using the
managed-by ref node of the bridge node and check for that in the operational data
store. That node is the connection node, so it must be there if device is connected
to the controller, otherwise configuration should fail, which is expected outcome.

Change-Id: I743269b5ca2f36f761bbb4a118a2837631e1847d
Signed-off-by: Anil Vishnoi <vishnoianil@gmail.com>
(cherry picked from commit ebc45c78ec38493dc954fc9d3a4a92f9266e9fde)