Updated git submodules
Project: mdsal master a6d0735732b202f40bd37f7ae291c8a4af9e3ccf

BUG-6028: check value types for encapsulation

This is a fast check to see if the argument matches expected
value class. If it does not it does not make sense to invoke
the serializer, as it will fail.

Change-Id: I4ca556fea057f2188fa1c34d303467565af38d5c
Signed-off-by: Robert Varga <rovarga@cisco.com>