Updated git submodules
Project: ovsdb stable/beryllium 3b76c4ab002be17843f6edaf6bdc3680ef3c30bc

Make SchemaVersionMismatchException logging consistent

The logging for the exception was not consistent: The log level varied,
the output varied and the logging of the stack trace varied.

Now a single utility method is called for the log. The level is set
to debug, same text, and the exception message rather than the trace
is logged. In all cases these are not errors because it is expected
that different schema versions may not have support for some columns.

Change-Id: I29ef9d30e2044483da1dfd9dc573d13715d222f1
Signed-off-by: Sam Hague <shague@redhat.com>