Updated git submodules
Project: lispflowmapping master d3fa81e9474ff4659e36c1bdb837556947ee4b6d

Merge "Add radix trie to HashMapDb for fast IP LPM"

Add radix trie to HashMapDb for fast IP LPM

This adds a radix trie implementation for fast IP longest prefix
matching leveraged for now in HashMapDb to speed up mappings lookup. The
change is not fully transparent as it introduces a number of ILispDAO
API and map-cache implementation updates.

An important caveat to consider is that authkey lookup logic in map-caches
cannot be changed to use the tries because keys are not always found
under the longest prefix match. This is a consequence of the fact that
we now accept generic auth keys for aggregate prefixes, i.e., we don't
have one-to-one matching between mapping and auth entries, and we store
both auth keys and mappings in the same map-cache.

Change-Id: I4e7d2badce3b3cd31030fbea87f7366ea6b41035
Signed-off-by: Florin Coras <fcoras@cisco.com>