Updated git submodules
Project: controller master 70b924258e6c87219c38bcfefb3781ee3d2e2d02

BUG-5903: do not rely on primary info on failure

This makes sure we check for failure before touching the result,
which is null if a failure occurs.

In order to keep disagnosti information we add a reference
to the message class being broadcast.

Change-Id: I26ab31a45916d11b61b990020bed89ae87233b14
Signed-off-by: Robert Varga <rovarga@cisco.com>