Updated git submodules
Project: ovsdb stable/beryllium 6610b077d29dfd7d29cc424fab652a7918865226

BUG 5945: Tunnel updates through Genius not working

When creating tunnels through Genius, bridge information is present only
in operational and not in config. So any data chages that come in will not
have any augmentation. This was fixed as part of BUG 4644 but changes for
DataTreeListener refactored the code and this scenario was missed.

Change-Id: Ie4982d887e555a40984a5f4ff1211b785fe5ee6d
Signed-off-by: Vishal Thapar <vishal.thapar@ericsson.com>