Updated git submodules
Project: ovsdb master 8d674cdadfd2aafb0605ab1de887d1bb44cdefad

BUG 5945: Tunnel updates through Genius not working

When creating tunnels through Genius, bridge information is present only
in operational and not in config. So any data chages that come in will not
have any augmentation. This was fixed as part of BUG 4644 but changes for
DataTreeListener refactored the code and this scenario was missed.

Change-Id: Ie4982d887e555a40984a5f4ff1211b785fe5ee6d
Signed-off-by: Vishal Thapar <vishal.thapar@ericsson.com>