Updated git submodules
Project: openflowplugin master fd27dce2179b4a7d82dee48563199ca104aa3fd2

Merge "Bug 5602: Fixed FlowRemoved notification"

Bug 5602: Fixed FlowRemoved notification

This patch fixed the condition on which FlowRemoved notification was
made. The FlowRemoved notification needs to be made even when an
instance of ItemLifecycleListener is not available. Therefore, this
patch changed the location to publish the notification in the code. This
change made the OF plugin publish the FlowRemoved notification to its
listeners even when the statistics manager is running.

Change-Id: I279643a118537d0e3bca43358520be15efba24e8
Signed-off-by: Hideyuki Tai <Hideyuki.Tai@necam.com>