Updated git submodules
Project: ovsdb master 261669c036e9f89b409faeaa6badb4b33ecc4468

BUG 5746 - Ovasb QoS and Queue enhancements part 2

Fixes 3 integration test errors.  2 errors due to bug
in the new Queue code.  The other error was in the integration
test code as a result of the model changes.

Also, added support to set the qos-id and queue-id based
on the presence of the 'opendaylight-qos-id' and
'opendaylight-qeueue-id' external-id values.  These
external-id keys were placed in the qos and queue
tables respectively by the original qos and queue
plugin support. The corresponding value was the
id value for the row entry used by the md-sal.

The updated qos and queue model enhancement code
use the 'opendaylight-iid' key and stores the
corresponding instance identifier for the qos or
queue object.  Support for using the previous
external-id keys provides backward compatiblity
when connecting to OVS hosts which contain qos
and queue entries created using the older plugin.

Change-Id: Ifd97b1f41a497cadd1060365122cae1fd7cbfe84
Signed-off-by: Eric Multanen <eric.w.multanen@intel.com>