Updated git submodules
Project: controller master c9d61ee66367d819319bb8ccfa9f9b0555264d86

BUG-5280: introduce request/response Envelope

This is a follow-up patch to move sequence information from
request/response structure and making it part of an Envelope,
which is allocated by the SequencedQueue.

Change-Id: I341118850d9c5835bab0b491f59b95264f31e5ef
Signed-off-by: Robert Varga <rovarga@cisco.com>


BUG-5280: implement message queueing

This patch implements the basic queueing and timeout retry mechanism
in ClientActorBehavior.

This implementation is not very efficient, as each send goes through
the actor's mailbox, but it gets the job done and is correct. It will
be optimized in a follow-up patch, which will refactor internal
workings so that SequencedQueue is fully thread-safe and correct with
regard to request enqueue, timeouts and retries.

Change-Id: I207a30877328dbdc08d42f76a0db55b5ae162de5
Signed-off-by: Robert Varga <rovarga@cisco.com>