Updated git submodules
Project: genius master 4ab8ad0e841c064788fb516ce702d3d2d3e54113

Fix Flow/GroupEntity & Co. toString() to be useful

Fixes a number of strictly speaking slightly separate problems: 

  1. FlowEntity/GroupEntity: The way the parent class super.toString()
was appended at the end instead of at the start was visually just very
ugly and confusing.  Removal of toString in AbstractSwitchEntity
accordingly, because .. that can never get called now.

  2. The use of the private internal "polish notation prefix" instead of
the classes public (as in getter and setter) field names  was just ugly
and confusing.

  3. ActionInfo, NxMatchInfo, MatchInfo, InstructionInfo and BucketInfo
were all simply completely missing a toString().  This is very bad
because they are contained in FlowEntity (GroupEntity), and thus need
toString as well.
  
  4. The use of Arrays.deepToString & Arrays.toString is because Guava
18 which ODL uses as of now does not do this automatically (so you get
"[J@493968a9" etc.).  With the switch to Guava 19 proposed in
https://git.opendaylight.org/gerrit/#/c/41515/ this will not be required
anymore.  It doesn't hurt either then, so for now to make this proposed
change not require that upgrade yet and keep things separate, we must
use it for all array (but not List) type fields.  FTR: The minor
disadvantage of Arrays.(deep)ToString is that it renders
omitNullValues() ineffective, so once we are successfully on Guava 19 on
master, it would leader to shorter = nicer toString() output to remove
them again.
  
This overall change should, of course, not have any impact anywhere,
because toString methods are ONLY supposed to be used for debugging and
logs; any code that relies on e.g. parsing their output would be a bug.

Change-Id: I942842c13670a7d9ce4adb19581cdf1cebf43087
Signed-off-by: Michael Vorburger <vorburger@redhat.com>