Updated git submodules
Project: genius master 0548efba5cf0bcaf70631daf100ec7af90ffb143

Merge "Making all these fields final and explicitly initialize to null"

Making all these fields final and explicitly initialize to null

This makes the intention of the code in these classes much clearer to
read (the idea of the API as-is seems to be that always only either of
the 3 is set).

Removal of setInstructionValues in InstructionInfo is only consistent
then.

Change-Id: I4a0c540df0a4526f6ff326b5386260e291085275
Signed-off-by: Michael Vorburger <vorburger@redhat.com>