Updated git submodules
Project: vpnservice master 952571bf03763403c0650da6db499c0c2a767d6d

Merge "Fixed IBgpManager by moving it, as it was in clearly wrong directory"

Fixed IBgpManager by moving it, as it was in clearly wrong directory

In Eclipse this turns into a lot of red blood in the workspace.

For mvn CLI this is probably just ignored, but I'm pretty sure this
can't actually be working properly the way you intended it to being
where it currently is - look at the built
target/bgpmanager-api-0.3.0-SNAPSHOT.jar and note how the IBgpManager is
inside a single folder named, literally
"org.opendaylight.bgpmanager.api" (which is clearly wrong) instead of in
sub-folders org/opendaylight/bgpmanager/api. If it does work, then it's
a coincidence (it's clearly wrong in Java), and should be moved as
proposed here.

Change-Id: Iaeca062f3c029e7e1f07fc4503969b24617e44e7
Signed-off-by: Michael Vorburger <vorburger@redhat.com>