Updated git submodules
Project: netvirt master 9eaee29f3257244c474a51b8087f63d9315e3ca5

Create physnet bridges+ports and patch to br-int

NB: This commit depends on changes in the ovsdb
project: https://git.opendaylight.org/gerrit/#/c/41946/

AND NOW ALSO:
https://git.opendaylight.org/gerrit/#/c/42693/

Assume "provider_mappings=physnet1:ABC". ABC will either be (in this
order) (1) the name of a port already configured on br-int, (2) the name
of a bridge already configured on OVS, or (3) the name of a physical
interface. So for each option:

(1) ABC is a port already on br-int
At OVSDB node add/update event:
  no op
When adding the ElanInterface to the ElanInstance:
  ABC is the name of the port, use that

(2) ABC is the name of a bridge on OVS
At OVSDB node add/update event:
  patch the bridge to br-int, br-int side port name is ABC-patch
When adding the ElanInterface to the ElanInstance:
  ABC-patch is the name of the port, use that

(3) ABC is the name of a NIC
At OVSDB node add/update event:
  add the port directly to br-int
When adding the ElanInterface to the ElanInstance:
  ABC is the name of the port, use that

Change-Id: Ibf6851d420c8197e11a6503b950fa381b829499c
Signed-off-by: Josh <jhershbe@redhat.com>