Updated git submodules
Project: sfc master c0ac2ee2b33839293465e26c69604588880f48c3

BUG 6268: service-function-dictionary grouping is redundant as it is

The grouping service-function-dictionary is already defined in the SFF
configuration with the same name. It was not desired to repeat the name
nor reuse the same fields, so additionally to change the name,
the following fields should be removed:

     list sff-interfaces
     leaf failmode

Additionally in this grouping, the grouping dpn-locator is used and
it should be renamed to dpn-grouping since the dpn-id defined inside
is not going to be defined as other locators.

Change-Id: Ib2042251ad6850da007979ad9dc861b4465b4a7d
Signed-off-by: danipeon <daniel.peon.quiros@ericsson.com>